Add files via upload
This project visualizes the latest Covid-19 India Status data, allowing a better understanding of the impact of the pandemic on India. The data includes the total number of confirmed cases, total number of active cases, total number of recovered cases and total number of deaths in India. The data is updated daily and provides a clear insight into the current Covid-19 situation in India. This project is a useful tool for understanding the impact of the pandemic in India and for making informed decisions about the health of its citizens.